feat(stdlib)!: improve the memory usage and the usability of `gradientCast`
The algorithm now selects a single path per node across the multiple paths that are locally available.
There is also no need to be informed if the environment is a Riemannian manifold, the metric is used to deduce it.
The new implementation uses exchange to share different paths with different neighbors.